requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pydantic==2.6.4
python-dotenv==1.0.1
requests==2.31.0
numpy<2
scikit-learn==1.4.2
google-generativeai==0.5.4
redis==5.0.4
sentence-transformers==2.6.1
openai
anthropic
mistralai